requirements.txt
opensearch-py==2.6.0
boto3==1.35.0
pydantic==2.9.2
pydantic-settings==2.6.1
tenacity==8.5.0
fastapi==0.115.4
uvicorn[standard]==0.32.0
langchain==0.2.16
langchain-openai==0.1.25
langchain-community==0.2.16
langgraph==0.2.0
openai==1.50.0
pandas==2.2.0
python-dotenv==1.0.1
pytest==7.4.3
PyJWT==2.9.0
httpx==0.27.0
cryptography==43.0.0